Add location to ctkVTKDataSetModel
When using both the cell data and the point data, one must be able to
know where the current array comes from to be able to appropriately use it
(for example with a vtkAssignAttribute filter).
This commits add the location, i.e. provenance of the array in the model's,
items along the corresponding functions for reading. No writing functions
were added as the location is completly related to the array itself.
Changing the array location probably doesn't make sense unless it's done
explicitly on the data.

This prompted a change on the ctkVTKDataSetArrayComboBox to add an read
methods on the current location.

The ctkVTKDataSetModel testing was modified accordingly.