BUG: listen for unique signals from the title text property
Previous fix set up the scalar bar widget to listen for the new
unique signals from the text property widget and emit a modfieid,
but only for the labels. This change adds that functionality for
the title text property.
The ctkVTKScalarBarWidget now will emit a modified() signal when
the color, opacity, font, or font style changes for the title.
This fixes the bug that the vtkScalarBarActor wasn't updating
on changes in the ctkVTKScalarBarWidget Title section.

Slicer issue #3027
http://www.na-mic.org/Bug/view.php?id=3027