Allow reload in case of external changes
Add ctkSettingsDialog::reloadSettings() method to allow reloading
settings in case of external changes. Rename / make public the
ctkSettingsPanel::updateProperties() method (to reloadSettings() also)
in order to implement this.

This provides a way to make the settings dialog / panels aware of
changes made to the relevant QSettings outside of the dialog, which is
both an obvious bug in its own right, but more importantly prevents the
dialog from overwriting such external changes (see commontk/CTK#450).