Improve Python access to ctkSettingsPanel
Add a Q_INVOKABLE overload of ctkSettingsPanel::registerProperty in
order to make the method usable from PythonQt. The overload takes the
signal as a QByteArray, to match the behavior of qt.SIGNAL, so that
users do not need to explicitly cast that to a str in order to call the
method. Also expose settings/setSettings as a Q_PROPERTY, for similar
reasons.