Update dashboard driver script to submit a regular project layout.
As discussed during the CTK hackfest, instead of submitting submitting
each CTK libraries, plugins, applications as a separate CDash subproject.
This commit reverts to a "regular" CDash layout where the
configure/build/test results are all presented on a single page. This will
be easier to maintain and less confusing for the developer.

Finally, the driver script is now downloaded directly from Github. This is
now possible by using CMake version >= 2.8.9 having built-in support
for https download. See [1]

Fixes #457

[1] http://cmake.org/gitweb?p=cmake.git;a=commit;h=4bcd84e65ad2f792c549989b9d773191ad75e5eb