Ensure ctk option are considered as superbuild candidate variable.
This fix issue introduced following commit dc70d03 and extend
the functionality of the external project dependency module by
introducing the "_SKIPCACHEVALUE" option.

By setting the variable <varname>_SKIPCACHEVALUE, the system
won't consider the cached value of an option. This is currently needed
in CTK because enabling top level option set the variable in the
current scope without setting the variable in the cache. For that
matter, the value in the current scope should be considered.

Note that the upstream project [1] hosting the external project dependency
macro hasn't yet been updated to support the skip cache value variable.

[1] https://github.com/commontk/Artichoke