Fix ctkSimplePythonShell build error supporting optional dependencies
This commit fix the build error and simplify how optional dependencies
associated with ctkSimplePythonShell are specified.

Since the 'target_libraries.cmake' file doesn't support the specification
of optional dependency. We introduce the variable <targetname>_OPTIONAL_DEPENDENCIES
that has to be set before the generation of DGraph input or build options
validation.

This variable allows to specify optional dependencies that will be
considered by both ctkFunctionGetIncludeDirs and ctkFunctionGetTargetLibraries.