Fix export of CTKTesting include directories
By leveraging the build system infrastructure and associating a
"target_libraries.cmake" file with CTKTesting, this commit ensures
that the associated directory are automatically exported
in CTKConfig.cmake

To successfully achieve this, the following has been done:
 (1) addition of an empty "target_libraries.cmake"
 (2) fix to DGraph so that the topological path of a standalone vertex
is the vertex itself and not the list of all vertices topologically sorted.
Note that more work would be required to teach DGraph to detect invalid
labels.
 (3) teach ctkFunctionCollectTargetLibraryNames how to handle empty
list of target libraries.