Add 'rules' to install from superbuild
Import ExternalProject_Install_CMake from commontk/Artichoke@b0cf6fc85d5a,
and use to install (some) external projects when building the install
target of a CTK superbuild.

Note that this (currently) does not install DCMTK, KWStyle or QtTesting
or PythonQt. KWStyle is probably not needed, but the other three should
be fixed (and also, the non-installed CTK build should use their build
trees rather than doing local installs). As a result, this is *NOT*
sufficient yet to produce a usable install of CTK from the superbuild.