Added static method to set multisampling.
This allows setting a static boolean variable in the ctkVTKAbstractView
class that determines whether or not future instances have
multisampling (FSAA) enabled by default. This is necessary because,
at least in some circumstances, the multisamples OpenGL property must
be set before the OpenGL context is created -- and this happens during
class initialization. Closes #463.