Remove warning output when double slider range is bound to double max
toInt was warning the developer that it failed to use ints to represent
the double value. This is to track issues with the single step.

In this case (ctkDoubleSlider::setRange(-DOUBLE_MAX, DOUBLE_MAX);), the
limitation does not come from the single step, it is simply expected to
fail and there is nothing we can do about it (expected behavior).

Therefore we do not print error messages.

See Slicer issue 3755