BUG: Prevent VTK render call in ctkVTKRenderView::setupRendering
This commit fixes error reported below by moving the initialization
of the orientation widget after disabling the rendering and setting up
the RenderEvent observer. This will prevent vtkRenderWindow::Render()
calls from happening when the context is not ready. See [1]

This was needed because setting the interactivity of the widget is only
possible if it is enabled and if it is enabled, it ends up calling Render()
on the interactor.

[1] http://www.vtk.org/Wiki/VTK/OpenGL_Errors#Apple_Mac_OSX

//------
invalid drawable

ERROR: In /path/to/Rendering/OpenGL/vtkOpenGLRenderer.cxx, line 1202
vtkOpenGLRenderer (0x7ff6f45225e0): failed after Clear 1 OpenGL errors detected
  0 : (1286) Invalid framebuffer operation
//------

Co-authored-by: Julien Finet <julien.finet@kitware.com>