BUG: Fixed missing list elements problem
When there are two strings with the only difference that one is longer (e.g., "Label Statistics" and "Label Statistics (BRAINS)") then one of them is missing from the search results.

The problem was that QAbstractItemModel::match was used for checking if the string is already in the model and by default Match uses Qt::MatchStartsWith option.

Fixed by using Qt::MatchFixedString.