BUG: Prevent infinite recursion in Qt log handler
Slicer sometimes crashed on exit because in Qt log handler in ctk::qtHandleString while executing line QTextStream s(&str)
a QDeviceClosedNotifier::QDeviceClosedNotifier() object was insantiated that tried to log a message, which eventually got to
ctk::qtHandleString, creating an infinite recursion.

To solve this, a counter is added that is incremented at each recursion level and when it reaches a threshold then it does not try
to handle the message, thus preventing the crash.