BUG: Fixed Console Shift+Home behavior
The console had an annoying bug: Shift+Home did not select text up to the start of line, just jumped to the start of line
(while Shift+End already properly selected text up to the end of line).

The problem was that Home key handling was overridden (to jump to the end of the prompt instead of start of line) but the Shift modifier was ignored.