Added file upload validation to saveImpl of ctkXnatFile
New class ctkXnatResourceCatalogXmlParser can parse the catalog xml of a
resource and returns the md5 hashes on the server. It is implemented as a
SAX parser.
The ctkXnatAPI's parseXmlResponse uses the new parser in order to process
the reply.
If the file validation fails an exception is thrown and the file is deleted.