Added refesh function to ctkXnatTreeModel
Therefor also a "force" flag was added to xnatObject::fetch, in order to
allow a fetch even if the object was already fetch previously.

This is necessary if the data was changed on the server e.g. via the XNAT
webinterface and hence is not consistant any more.

In order to update the treemodel accordingly, a refresh function was added
there as well, which calls fetch recursively on all ctkXnatObject that are
already fetched.