ENH: Give more information for not showing the message box again
This change was inspired by using a ctkMessageBox with three options
and wanting to be clear what would happen if the don't show again
option was used. Add the text from the button with the accept role
to the text string so that the user knows what will happen
by default if they chose not to show this message box again.