Update PythonQt to patched-5, which now includes selected updates from patched-6
This commit integrates recent fixes from the PythonQt master, like in the
patched-6 branch, but skips a commit that changes the generated Python Qt API.

The patched-6 branch includes the contents of patched-5 plus PythonQt commits up
to and including commit http://sourceforge.net/p/pythonqt/code/403/.  Commit
http://sourceforge.net/p/pythonqt/code/397/ effectively changes the generated
Python Qt API by removing the singleShot attribute on QTimer to avoid a name
conflict with the static method of the same name. For example, code like this
would need to be rewritten to use the setSingleShot() method instead of setting
the singleShot attribute:

myTimer = qt.QTimer()
myTimer.singleShot = True

The process for updating to patched-6 while maintaining backwards compatibility
is planned to be addressed separately.