ENH: add the capability to export DICOM to disk
Add the export series, study, patient option to the right click menus.
Gets the files associated with a series and constructs
an export path based on the selected directory, using the patient
ID and name, study date and description, series number and
description.
The destination file is named sequentially, so a series will
be named 000000.dcm through to n-1.dcm.
Call the export series method for each selected study.
Call the export study for each selected patient.
Use a progress dialog to give the user feedback on the
status of the DICOM export.
Use message box pop ups to report errors to the user.
Fix a typo in a method name:
confirmDeleteSelectedIUDs -> confirmDeleteSelectedUIDs

Slicer Issue #3163