Remove superfluos test on fontconfig
The test is run but only used to see if vtkRenderingFreeTypeFontConfig is
to be linked, but it is already tested whether the vtk target
vtkRenderingFreeTypeFontConfig actually exists.

In addition on Debian the FontConficConfig.cmake file doesn't extist, so the
test fails and linking as well, because the needed
vtkRenderingFreeTypeFontConfig is not addded to the linker line.