CTKCore: Generalize test checking if dl library is available.
This commit generalizes the fix integrated in 9684161(fix Core
linking: FreeBSD does not have a libdl)

FreeBSD documentation also confirms that there are no dl library on
that system. It means checking for the presence of the library is a valid
approach. See http://www.freebsd.org/cgi/man.cgi?query=dlopen