ENH: ctkPytonConsole: Move cursor between parentheses if parameters detected
It improves the ctk python console so that the cursor is
automatically moved between the parentheses for a function
when autocompleted with TAB (this is only done when the
funtion has arguments).

This commit is based on commit 42b84f1 originally associated with
topic python-autocomplete-parameter-detection from Christopher
Mullins.

It has been rebased and adapted by Mayeul Chassagnard.

It fixes issues reported in http://na-mic.org/Mantis/view.php?id=1227

Co-authored-by: Mayeul Chassagnard <mayeul.chassagnard@kitware.com>
Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>