BUG: ctkPythonConsole: Fix autocompletion issue
This commit fixes Slicer issue 4236.

For example the issue happened after typing this:

  import SampleData
  SampleData.SampleDataLogic().dow

and trying to autocomplete with 'Tab'. The autocompletion was working
only if we removed the '()' but in the command you must have the '()'
to make the command works. Then, we should leave the '()' for the
command shell but removed it to search the autocompletion.

More specifically it implements the following changes:

- Don't leave anymore the word we typed when autocompleting in
  case of there is something after

- Don't erase anymore the parentheses when autocomplete

- Leave the cursor at the end of the word completed (and don't put it at the very end).
  Moreover it puts the cursor between the parentheses of the world completed.

For more details, see http://na-mic.org/Mantis/view.php?id=4236