ENH: ctkAbstractPythonManager: pythonAttributes can instantiate class and return instance members
This function now allows to return all the attributes of an object,
even if it's a callable object: it instantiates it and returns instance members.

Moreover, it goes through all pythonAttributeName and iteratively can
access every attributes in one line.

For example, we can access to bar_instance_member
(according to the prerequisite file pythonAttributes-test.py), doing:
d.foo_class().instantiate_bar().[TAB]

Notes:

(1) An attribute of main named '_ctkAbstractPythonManager_autocomplete_tmp'
is created temporary during the research of auocompletion but immediately deleted

(2) Add a function dir_object returning the attributes of a PyObject into a StringList

(3) Can't yet instantiate a class with argument