ENH: Preserve actor order in vtkLightBoxRendererManager
The order of how 2D actors are added to the renderer greatly influences the rendering result. Therefore, for consistent rendering results actor ordering in slice viewer has to be preserved when just changing the input image data.

Old behavior: When input to slice viewer was set to NULL (no image is selected for display), the image actor was removed from the renderer. When an image was selected again, the image actor was added again. This changed the actor order.

Implemented fix: Modified the behavior to only hide the image actor when input is set to NULL. This way when image input is later set to non-NULL, the image actor does not get promoted to the top of the actor list (but keeps its original position).