COMP: If defined, pass CMAKE_TRY_COMPILE_CONFIGURATION to inner build
This is required to ensure the binaries compiled using "try_compile"
and executed during the configuration step can find the Release runtime
libraries when no Debug runtime libraries are available.

By default, binaries compiler using "try_compile" are build in Debug.

This applies when building the project using "Microsoft Visual C++
compiler for Python 2.7", in that case only the Release runtime libraries
are installed as side-by-side assemblies in the C:/Windows/winsxs directory.