ENH: Improve ctkFittedTextBrowser collapsed text formatting
Show ellipsis (...) where the collapsed text is cut.

Remove teaser separator(line break) character instead of replacing with space. This allows better control of what is displayed.
For example: "Some teaser<br>. This is more info." text expanded version was displayed as "Some teaser . This is more info.", but
now it is shown as "Some teaser. This is more info."