BUG: Fixed multi-value argument parsing
When a multi-value argument was passed to the command-line parser, all subsequent parameters were considered part of the argument.
This caused several test errors in Slicer: --ptyhon-script parameter was ignored if --additional-module-paths argument was specified before,
which made --python-script parameter ignored, and as a result, the provided Python script was not executed, which caused test failures
(py_nomainwindow_SlicerTestingExitFailureTest, py_nomainwindow_SlicerUnitTestWithErrorsTest, py_SlicerTestingExitFailureTest).

The root cause of the problem was that the check that compared a parameter to a known argument was incorrect: the parameter name
contained the prefix (--python-script) and this string was searched among argument names (such as python-script).

Fixed by parsing each parameter after a multi-value argument. If a known argument is found, it is not added to the multi-value argument.

Added test: modified ctkCommandLineParserTest1 to test that arguments specified after a multi-value argument are parsed correctly.