ENH: Fix ctkMenuComboBox behavior in Qt5
In Qt5, when QCompleter sends its activated() signal, QComboBox sets
its current index to the activated item, if found. Work around that behavior
by re-selecting the original item.

Additionally, ensure that an item is not added to the combobox when the
user enters a new string.

Together, these changes make ctkMenuComboBox behave like in Qt4.