BUG: Fix ctkDoubleSpinBox memory corruption
The ctkDoubleSpinBoxPrivate instance is used by both ctkDoubleSpinBox and the
ctkQDoubleSpinBox that the private instance creates. The private instance adds
the ctkQDoubleSpinBox instance as a child of the ctkDoubleSpinBox instance.
This commit moves ownership of the ctkDoubleSpinBoxPrivate instance from
ctkDoubleSpinBox to ctkQDoubleSpinBox to ensure a destruction order that doesn't
cause memory corruption.