ENH: Make ctkSettingsTest1 more robust
ctkSettingsTest1 tests setting and restoring the main window position. If the
system running the tests had a decoration--such as dock/menu bar/task bar--in
the hard-coded position of the window, then the test would fail. This happened
because the logic that restores the window position ensures that the window
position avoids those decorations.

Update ctkSettingsTest1 to account for the available desktop geometry when
positioning the test window.