ENH: Fix ctkErrorLogQtMessageHandlerWithThreadsTest1 with Qt5
Fix ctkErrorLogQtMessageHandlerWithThreadsTest1 with Qt5.

The message handler that ctkErrorLogQtMessageHandler installs to capture error
messages--ctkErrorLogModelQtMessageOutput--discards messages if it detects a
recursion depth greater than 10. However, the recursion count doesn't
differentiate between actual recursion and concurrent execution of the handler
in different threads. ctkErrorLogQtMessageHandlerWithThreadsTest1 starts 15
threads and fails because the recursion check discards error messages.

To work around the problem for now, reduce the number of threads in the test.