ENH: Speed up ctkDICOMBrowser removing processEvents call
This commit removes call to "processEvents()" initially added
in cce76a9 (New ctkDICOMBrowser including a new example application
ctkDICOM2).

This call added extra overhead and was causing a x2 slowdown when
importing large collection of DICOMs. In our example, importing
556 files was taking ~8s and without this function call, time was reduced
to 3.5s.

We also observed that canceling the import process is still possible.