Add ctkVTKDiscretizableColorTransferWidget
Add a new widget to preview and edit a color transfer function given as
a vtkScalarsToColors.
The widget owns an instance of vtkScalarsToColorsContextItem that can be
added to a vtkContextView. vtkScalarsToColorsContextItem holds three
different charts to represent the information associated to the color transfer
function :
  - vtkDiscretizableColorTransferFunctionChart: Uses
  vtkControlPointsItem to edit the color transfer function

  - vtkScalarsToColorsPreviewChart: Preview the current color transfer
  function.

  -vtkScalarsToColorsHistogramChart: Display the histogram of an input
  image while applying the color transfer function to the histogram.

The widget also owns an instance of vtkScalarsToColorsComboBox to access
different presets.