BUG: Re-add range reset on bounds change if range is invalid
The setAxesToChartBounds call has recently been removed so that the chart extent is not reset when the transfer function bounds change. This prevented resetting the extent defined by the user when shifting the transfer function. However, if there was no valid extent defined, without this call the chart was not shown at all. By resetting the extent only if it was invalid solves both issues.