BUG: Fixed failed re-import of DICOM images
IDs of last inserted patient, study, series items were cached in the ctkDICOMDatabase to avoid re-inserting when adding each file instance. However, if this patient, study, or series are removed from the database, then their IDs have to be cleared to force adding them back to the database when adding a file instance. Clearing of IDs when items are deleted has been implemented in ctkDICOMDatabase. The problems is that an application may use multiple database connections (ctkDICOMDatabase instances) to connect to the same database, and delete items corresponding to the cached IDs.

Fixed the problem by deleting cached item IDs once, just before a batch of instances are inserted to the database.

To define what is the beginning and end of a "batch", startIndexing and endIndexing method was added to ctkDICOMIndexer. They are called automatically when any of the ctkDICOMIndexer::add...() methods is used. Start/end indexing methods are exposed on the public API as well so that applications can do multiple ctkDICOMIndexer::add...() calls in one batch.

Also fixed/completed Python wrapping for ctkDICOMTableManager, ctkDICOMBrowser, and ctkDICOMDatabase.

Fixes https://issues.slicer.org/view.php?id=3806