COMP: Support building against VTK version introducing QVTKOpenGLNativeWidget
Related Slicer issue https://issues.slicer.org/view.php?id=4598

Following VTK commit kitware/VTK@45704131f, QVTKOpenGLWidget was renamed
to QVTKOpenGLSimpleWidget and the QVTKOpenGLWidget class was updated to be
able to work with a  Qt OpenGL context (though the use of a new class named
QVTKOpenGLWindow).

These changes were initially motivated to support creating stereo-capable
applications that needs multiple framebuffers. The updated class also provides
a better management of OpenGL resources (check for valid resources, reduces
flickering ...) and also adds support for HiDPI. But there are some gotchas
discussed below.

Then, following VTK commit kitware/VTK@fd9cdefeb, QVTKOpenGLSimpleWidget
was renamed to QVTKOpenGLNativeWidget


Why not keep using QVTKOpenGLWidget ?
-------------------------------------

It turns out that using the "new" QVTKOpenGLWidget is not possible when
it is the child of a QScrollArea or QMdiArea. In that case, the widget
will be changed into a "Native" one, and this is not supported.

Waiting the QVTKOpenGLWidget is improved to gracefully handle the different
use cases, this commit simply make use of QVTKOpenGLNativeWidget when it
is available.


Note about "black" screen issue
-------------------------------

An issue impacting using of older QVTKOpenGLWidget or newer QVTKOpenGLSimpleWidget
was fixed in 1c33914aa (Fix issue where picking happened with a bad context).
This was causing the entire application to turn black when the user
attempted to using picking in a render window. Regression originally
introduced in kitware/VTK@1eb083a80 (use hardwareselector to perform picking)

See https://gitlab.kitware.com/vtk/vtk/issues/17370


Which version of VTK are supported ?
------------------------------------

The renaming of QVTKOpenGLSimpleWidget to QVTKOpenGLNativeWidget and the
fix for the "back screen" issue means that using VTK version between
kitware/VTK@1eb083a80~1 and kitware/VTK@fd9cdefeb~1 will most result
in broken CTK build or improperly working application.