STYLE: Simplify code introducing "ctkVTKOpenGLNativeWidget.h"
Since using typedef or #ifdef prevents the real class associated with
ctkVTKOpenGLNativeWidget from being wrapped in PythonQt, this commit creates
a derived class.

Note also that the use of Q_DECLARE_METATYPE and qRegisterMetaType was
considered but it revealed not to be a possible alternative.
See https://github.com/commontk/CTK/pull/828

Fixes #829

Co-authored-by: Andras Lasso <lasso@queensu.ca>