Fix QtDesigner crash in debug mode
Debug-mode Qt Designer crashed when tried to instantiate a ctkMatrixWidget because of an assert failure.
It may be useful to know that some operations fail, but making the Qt Designer crash is a too strong way of notifying developers (especially when developers only use CTK and not intend to fix its potential internal errors).

Fixed by adding a new CTK_SOFT_ASSERT macro that just logs a warning instead of terminating the application (and does it both in debug and release mode, to make application behavior consistent). It should be possible to use this macro instead of Q_ASSSERT where crashing the application in unexpected situations in debug mode should be avoided.