Add ctkErrorLogAbstractModel to CTKCore library
This commit was motivated to support use of the log model from
qSlicerCoreApplication. This is needed to ensure messages are appended
to the log file during the unloading of modules following the call to
"factoryManager()->unloadModules()" from within the qSlicerCoreApplicationPrivate
destructor.

It is a follow up of f49c19a82 (move Widget dependent classes from Core
to Widgets. See #278)