BUG: Fix handling of PythonLibs
The FindPythonLibs module reads PYTHON_LIBRARY as a hint but it
also sets the same variable when successful. In the latter case,
the variable may be set to a list containing target_link_libraries
keywords like "optimized" and "debug" instead of a single filepath.
When read as a hint, though, FindPythonLibs only understands a
single filepath. Repeated unguarded calls to the module may result
in a faulty value of PYTHON_LIBRARY like:

  "optimized;optimized;optimized;python3.lib;debug;debug;python3d.lib"

Such values entail errors when the linker tries to link against
the keywords "optimized" or "debug".

This commit implements the following changes:

* Fix usage of FindPythonLibs result variables

  The documented output variable of FindPythonLibs is
  PYTHON_LIBRARIES instead of PYTHON_LIBRARY, which is an old
  relict from pre CMake 3.0 (minimal required CMake version of CTK).

* Fix linking to Python on Windows in Debug configuration

  Linking to a debug version of Python has many implications and is
  generally quirky enough that many famous projects like Boost and
  Qt decided to link against the release version of Python instead.

  Handling of PYTHON_LIBRARIES/PYTHON_LIBRARY was refactored into
  ctkFunctionExtractOptimizedLibrary.

  For non-MSVC/non-multiconfig build systems it is basically a NO-OP
  and does not change behavior.

Signed-off-by: Stefan Dinkelacker <s.dinkelacker@dkfz-heidelberg.de>